PR-15 Rename some classes and delete comments
